break out CleanedIdentitySerializer
